Correct usage of PSD for LOSC GW analysis whitening.
Ensured that whitening of strain is done using the same PSD that is used for matched. Note that in the LOSC tutorial, they use two different PSDs for the task of whitening strain and for matched filtering, so I have left the option to use the same PSD or the two separate PSDs.